+++ app/model/ORM/Models/ModelEvent.php
@@ -114,7 +114,7 @@ class ModelEvent extends AbstractModelSingle implements IResource, IContestRefer
      * @return GroupedSelection
      */
     public function getScheduleGroups(): GroupedSelection {
-        return $this->related(DbNames::TAB_SCHEDULE_GROUP,'event_id');
+        return $this->related(DbNames::TAB_SCHEDULE_GROUP, 'event_id');
     }
 
     /**
@@ -139,8 +139,8 @@ class ModelEvent extends AbstractModelSingle implements IResource, IContestRefer
             'eventId' => $this->event_id,
             'year' => $this->year,
             'eventYear' => $this->event_year,
-            'begin' => $this->begin->format('c'),
-            'end' => $this->end->format('c'),
+            'begin' => $this->begin ? $this->begin->format('c') : null,
+            'end' => $this->end ? $this->end->format('c') : null,
             'registration_begin' => $this->registration_begin->format('c'),
             'registration_end' => $this->registration_end->format('c'),
             'name' => $this->name,